Renamed NSD client to more descriptive and inforamtive name
