+++ src/NsdPublicationChannelRegistryClient.js
